pyproject.toml
[project]
name = "llm-mafia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.0",
    "streamlit>=1.40.0",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]
